refactor: vitest mock setup
