+++ client/vitest.config.mjs
@@ -15,6 +15,6 @@ export default defineConfig({
       "@": path.resolve(__dirname, "./"),
       app: path.resolve(__dirname, "./app"),
     },
-    setupFiles: ["./tests/setup.ts"],
+    setupFiles: ["./tests/setup/global.ts"],
   },
 });
